判断删除

+++ mingrui-shop-service/mingrui-shop-service-xx/src/main/java/com/baidu/shop/service/impl/SpecificationServiceImpl.java
@@ -3,10 +3,14 @@ package com.baidu.shop.service.impl;
 import com.baidu.shop.base.BaseApiService;
 import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.SpecGroupDTO;
+import com.baidu.shop.dto.SpecParamDTO;
 import com.baidu.shop.entity.CategoryEntity;
 import com.baidu.shop.entity.SpecGroupEntity;
+import com.baidu.shop.entity.SpecParamEntity;
 import com.baidu.shop.mapper.SpecGroupMapper;
+import com.baidu.shop.mapper.SpecParamMapper;
 import com.baidu.shop.service.SpecificationService;
+import com.baidu.shop.status.HTTPStatus;
 import com.baidu.shop.utils.BaiduBeanUtil;
 import com.baidu.shop.utils.ObjectUtil;
 import com.google.gson.JsonObject;
@@ -15,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RestController;
 import tk.mybatis.mapper.entity.Example;
 
+import javax.annotation.Resource;
 import java.util.List;
 
 /**
@@ -28,15 +33,18 @@ import java.util.List;
 @RestController
 public class SpecificationServiceImpl extends BaseApiService implements SpecificationService {
 
-    @Autowired
+    @Resource
     private SpecGroupMapper specGroupMapper;
+
+    @Resource
+    private SpecParamMapper specParamMapper;
 
     @Override
     public Result<List<SpecGroupEntity>> getSepcGroupInfo(SpecGroupDTO specGroupDTO) {
         Example example = new Example(SpecGroupEntity.class);
         if (ObjectUtil.isNotNull(specGroupDTO.getCid()))
-        example.createCriteria().andEqualTo("cid",
-                BaiduBeanUtil.copyProperties(specGroupDTO,SpecGroupEntity.class).getCid());
+            example.createCriteria().andEqualTo("cid",
+                    BaiduBeanUtil.copyProperties(specGroupDTO,SpecGroupEntity.class).getCid());
         List<SpecGroupEntity> specGroupEntities = specGroupMapper.selectByExample(example);
         return this.setResultSuccess(specGroupEntities);
     }
@@ -59,7 +67,51 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
     @Transactional
     @Override
     public Result<JsonObject> deleteSpecGroup(Integer id) {
+
+        //删除规格组之前需要先判断一下当前规格组下是否有规格参数
+        //true : 不能被删除
+        //false -->
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",id);
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+        //System.out.println(specParamMapper.selectCount(example));
+        if (specParamEntities.size()!=0) return this.setResultError(HTTPStatus.OPERATION_ERROR,"当前规格组下有参数，请先删除规格组下的内容");
+
         specGroupMapper.deleteByPrimaryKey(id);
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<List<SpecParamEntity>> getSepcParamInfo(SpecParamDTO specParamDTO) {
+        SpecParamEntity specParamEntity = BaiduBeanUtil.copyProperties(specParamDTO, SpecParamEntity.class);
+
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",specParamEntity.getGroupId());
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+        return this.setResultSuccess(specParamEntities);
+    }
+
+    @Transactional
+    @Override
+    public Result<JsonObject> saveSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.insertSelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Transactional
+    @Override
+    public Result<JsonObject> editSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.updateByPrimaryKeySelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Transactional
+    @Override
+    public Result<JsonObject> deleteSpecParam(Integer id) {
+        specParamMapper.deleteByPrimaryKey(id);
+
+
+
         return this.setResultSuccess();
     }
 }
